LayoutTests/legacy-animation-engine/transitions/transition-to-from-auto-expected.txt

CONSOLE MESSAGE: line 371: Failed to pause 'left' transition on element 'test1'
CONSOLE MESSAGE: line 371: Failed to pause 'left' transition on element 'test2'
PASS - "left" property for "test1" element at 1s saw something close to: 0
PASS - "left" property for "test2" element at 1s saw something close to: 100
PASS - "left" property for "test3" element at 1s saw something close to: 50
